auth

+++ project/src/controllers/teams/destroy.php
@@ -6,13 +6,13 @@ use Core\Database;
 
 $db = App::resolve(Database::class);
 
-$currentUserId = 1;
+$currentUserId = $_SESSION['user']['id'];
 
 $team = $db->query('select * from teams where id = :id', [
     'id' => $_POST['id']
 ])->findOrAbort();
 
-//authorize($note['user_id'] === $currentUserId);
+authorize($team['user_id'] === $currentUserId);
 
 $db->query('delete from teams where id = :id', [
     'id' => $_POST['id']
